+++ txOutputs.go
@@ -1,16 +1,14 @@
 package psl
 
 type SimpleOutput struct {
-  From   string `cbor:"0,keyasint" json:"from"`
-  To     string `cbor:"1,keyasint" json:"to"`
-  Asset  string `cbor:"2,keyasint" json:"asset"`
-  Amount uint   `cbor:"3,keyasint" json:"amount"`
+  To     string `cbor:"0,keyasint" json:"to"`
+  Asset  string `cbor:"1,keyasint" json:"asset"`
+  Amount uint   `cbor:"2,keyasint" json:"amount"`
 }
 
 type MultiAssetOutput struct {
-  From   string        `cbor:"0,keyasint" json:"from"`
-  To     string        `cbor:"1,keyasint" json:"to"`
-  Assets []AssetOutput `cbor:"2,keyasint,toarray" json:"asset"`
+  To     string        `cbor:"0,keyasint" json:"to"`
+  Assets []AssetOutput `cbor:"1,keyasint,toarray" json:"asset"`
 }
 
 type AssetOutput struct {
@@ -19,9 +17,8 @@ type AssetOutput struct {
 }
 
 type MultiAddrOutput struct {
-  From   string       `cbor:"0,keyasint" json:"from"`
-  To     []AddrOutput `cbor:"1,keyasint,toarray" json:"to"`
-  Asset  string       `cbor:"2,keyasint" json:"asset"`
+  To     []AddrOutput `cbor:"0,keyasint,toarray" json:"to"`
+  Asset  string       `cbor:"1,keyasint" json:"asset"`
 }
 
 type AddrOutput struct {